% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemakerruntime_package.R
\docType{package}
\name{paws.sagemakerruntime-package}
\alias{paws.sagemakerruntime}
\alias{paws.sagemakerruntime-package}
\title{paws.sagemakerruntime: Amazon SageMaker Runtime}
\description{
The Amazon SageMaker runtime API.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
